refactor: extract prompt-building into own package
With this other commands such as 'delete' can make use of the same logic
to be able to build prompts.
